Created dataloader which rescaled batches
